Add Excitation Coil
Bugs:
Can place buttons/redstone on top/sides
Bounding Box is not edited.

Because current CF groovy is broken, this includes a build of the latest github. If you don't trust me, go build your own jar.

+++ overrides/groovy/classes/blocks/DustBlock.groovy
@@ -10,8 +10,8 @@ class DustBlock extends BlockFalling {
 		if (tab != null) {
             setCreativeTab(tab)
         }
-        setHardness(0.4 as float)
-        setResistance(0.4 as float)
+        setHardness(0.4F)
+        setResistance(0.4F)
         setHarvestLevel("shovel", 0)
 	}
 }
